Combined the search and filter feature into listclubs.jsx
listclubsfilter.jsx no longer used
