+++ app/imports/ui/components/ClubCard.jsx
@@ -91,7 +91,7 @@ class ClubCard extends React.Component {
               style={ { backgroundColor: `${labelColor(data, clubTypes)}`, color: 'white' } }
               as={ NavLink }
               exact
-              to={`/clubtype/${data}`}
+              to={`/clubs/${data}`}
             >
               {data}
             </Label>)}
